+++ Pspcl/Pspcl.Web/Controllers/ReportController.cs
@@ -16,10 +16,12 @@ namespace Pspcl.Web.Controllers
     {
         private readonly ApplicationDbContext _context;
         private readonly IStockService _stockService;
+        private readonly IBlobStorageService _blobStorageService;
         private readonly ILogger<DbInitializer> _logger;
-        public ReportController(ApplicationDbContext _dbContext, IStockService stockService, ILogger<DbInitializer> logger) {
+        public ReportController(ApplicationDbContext _dbContext, IBlobStorageService blobStorageService, IStockService stockService, ILogger<DbInitializer> logger) {
             _context = _dbContext;
             _stockService = stockService;
+            _blobStorageService = blobStorageService;
             _logger = logger;
         }
         
@@ -97,16 +99,16 @@ namespace Pspcl.Web.Controllers
         {
             try
             {
-                var filteredstockOutModels = _stockService.GetStockOutModels();
+                var filteredStockOutModels = _stockService.GetStockOutModels();
 
                 if (fromDate.HasValue && toDate.HasValue)
                 {
-                    filteredstockOutModels = filteredstockOutModels.Where(s => s.CurrentDate.Date >= fromDate.Value.Date && s.CurrentDate.Date <= toDate.Value.Date)
+                    filteredStockOutModels = filteredStockOutModels.Where(s => s.CurrentDate.Date >= fromDate.Value.Date && s.CurrentDate.Date <= toDate.Value.Date)
                          .OrderByDescending(s => s.CurrentDate)
                          .ToList();
                 }
 
-                return Json(filteredstockOutModels);
+                return Json(filteredStockOutModels);
             }
             catch (Exception ex)
             {
@@ -142,7 +144,7 @@ namespace Pspcl.Web.Controllers
         
         public JsonResult DownloadImage(string filename)
         {
-            string downloadStatus = _stockService.DownloadFileFromBlob(filename);
+            string downloadStatus = _blobStorageService.DownloadFileFromBlob(filename);
 
             if (downloadStatus == "DownloadFailed")
             {
